Added setting of score text Created Game.StartGame Set score to 0 in Game.StartGame Added GameController GameObject Added menu to GameController Moved in game esc key check to object that creates and updates the Game object Created a new Game object in the GameController.Start Now call Game.Update from GameController.Update

+++ Marble Motion/Assets/Scripts/MainMenuController.cs
@@ -13,7 +13,7 @@ public class MainMenuController
     }
 
     // Set the score text
-    public void SetScoreText(int score)
+    public void SetScoreText(long score)
     {
         scoreController.SetScoreText(score.ToString());
     }
